backend/pyproject.toml
[project]
name = "webapp-coolbeans"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.123.9",
    "fastapi-cloud-cli>=0.6.0",
    "mypy==1.19.0",
    "pytest>=9.0.2",
    "roslibpy>=2.0.0",
    "psycopg2-binary>=2.9.11",
    "types-psycopg2>=2.9.21.20251012",
    "python-dotenv>=1.2.1",
]
